pyproject.toml
[tool.poetry]
name = "halite-tournament"
version = "0.1.0"
description = ""
authors = ["Nick Malaguti <nmalaguti@users.noreply.github.com>"]
readme = "README.md"
packages = [
    {include = "project", from = "src"},
    {include = "tournament", from = "src"},
]

[tool.poetry.dependencies]
python = "^3.11"
trueskill = "^0.4.5"
mpmath = "^1.3.0"
django = "^4.2.2"
gunicorn = "^20.1.0"
social-auth-app-django = "^5.2.0"
whitenoise = "^6.5.0"
django-storages = {extras = ["boto3"], version = "^1.13.2"}
django-configurations = {extras = ["database"], version = "^2.4.1"}
psycopg = {extras = ["binary", "pool"], version = "^3.1.9"}
sentry-sdk = "^1.26.0"
django-crispy-forms = "^2.0"
crispy-bootstrap3 = "^2022.1"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.3"
pygithub = "^1.59.0"
mashumaro = "^3.8.1"
python-dxf = "^11.1.1"
docker-image-py = "^0.1.12"
tenacity = "^8.2.2"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
ipython = "^8.14.0"
ptpython = "^3.0.23"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
sections = ['FUTURE', 'STDLIB', 'DJANGO', 'THIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER']
known_django = ['django', 'rest_framework']
